+++ lib/store/contact.dart
@@ -189,12 +189,16 @@ class ContactRepository implements AbstractContactRepository {
 
   @override
   Future<void> deleteContact(ChatContactId id) async {
-    final HiveRxChatContact? oldChatContact = contacts.remove(id);
+    final bool isFavorite = favorites.containsKey(id);
+
+    final HiveRxChatContact? oldChatContact =
+        (isFavorite ? favorites : contacts).remove(id);
 
     try {
       await _graphQlProvider.deleteChatContact(id);
     } catch (_) {
-      contacts.addIf(oldChatContact != null, id, oldChatContact!);
+      (isFavorite ? favorites : contacts)
+          .addIf(oldChatContact != null, id, oldChatContact!);
       rethrow;
     }
   }