+++ lib/util/obs/rxlist.dart
@@ -178,6 +178,12 @@ class RxObsList<E> extends ListMixin<E>
   }
 
   @override
+  void removeRange(int start, int end) {
+    _value.removeRange(start, end);
+    refresh();
+  }
+
+  @override
   void clear() {
     _value.clear();
     refresh();